chore: remove the former graphql API
