+++ unfollow-ninja-server/package.json
@@ -57,10 +57,8 @@
     "@types/oauth": "^0.9.1",
     "@types/p-limit": "^2.2.0",
     "@types/twit": "^2.2.28",
-    "apollo-server": "^2.23.0",
     "big-integer": "^1.6.48",
     "dotenv": "^8.2.0",
-    "graphql": "^15.5.0",
     "hot-shots": "^8.3.1",
     "i18n": "^0.13.2",
     "ioredis": "^4.27.1",
