+++ unfollow-ninja-server/src/api.ts
@@ -1,86 +1,73 @@
 import 'dotenv/config';
 
 import * as Sentry from '@sentry/node';
-import kue from 'kue';
-import { ApolloServer, AuthenticationError } from 'apollo-server';
-
-import resolvers from './api/resolvers';
-import { typeDefs } from './api/schema';
-import logger, { setLoggerPrefix } from './utils/logger';
+import Koa from 'koa';
+import Router from 'koa-router';
+import session from 'koa-session';
 import Dao from './dao/dao';
+import kue from 'kue';
+import logger, {setLoggerPrefix} from './utils/logger';
+import { createAuthRouter } from './api/auth';
+
+function assertEnvVariable(name: string) {
+  if (typeof process.env[name] === 'undefined') {
+    logger.error(`Some required environment variables are missing (${name}).`);
+    logger.error('Make sure you added them in a .env file in you cwd or that you defined them.');
+    process.exit();
+  }
+}
+assertEnvVariable('WEB_URL');
+assertEnvVariable('COOKIE_SIGNING_KEY');
 
 setLoggerPrefix('api');
 
 const SENTRY_DSN = process.env.SENTRY_DSN_API || undefined;
 if (SENTRY_DSN) {
-    Sentry.init({ dsn: SENTRY_DSN });
+  Sentry.init({ dsn: SENTRY_DSN });
 }
 
+assertEnvVariable('REDIS_URI');
+assertEnvVariable('POSTGRES_URI');
 const dao = new Dao();
 const queue = kue.createQueue({redis: process.env.REDIS_KUE_URI});
+const authRouter = createAuthRouter(dao, queue);
 
-const context = async ({req}) => {
-    const ip = req.headers['x-forwarded-for'] || req.connection.remoteAddress;
-    logger.info(ip + ' - ' + JSON.stringify(req.body.query, null, null));
-    const { uid } = req.headers;
-    if (!uid) {
-        throw new AuthenticationError('uid required');
-    }
-    const session = await dao.getSession(uid);
-    const setSession = (data) => dao.setSession(uid, data);
-    return { uid, session, setSession, dao, queue, req };
-};
+const router = new Router()
+  .get('/', ctx => {
+    ctx.body = {status: 'ᕕ( ᐛ )ᕗ Hello, fellow human'};
+  })
+  .use('/auth', authRouter.routes(), authRouter.allowedMethods())
+  .get('/test', ctx => {
+    ctx.body = {status: ctx.session};
+  });
 
-// sentry integration
-// from https://gist.github.com/nodkz/d14b236d67251d2df5674cb446843732
-const apolloServerSentryPlugin = SENTRY_DSN ? {
-    requestDidStart() {
-        return {
-            didEncounterErrors(rc) {
-                Sentry.withScope((scope) => {
-                    scope.addEventProcessor((event) =>
-                        Sentry.Handlers.parseRequest(event, (rc.context as any).req)
-                    );
-
-                    // username
-                    const username = (rc.context as any).session?.user?.username;
-                    if (username) {
-                        scope.setUser({
-                            ip_address: (rc.context as any).req?.ip,
-                            username,
-                        });
-                    }
-
-                    scope.setTags({
-                        graphql: rc.operation?.operation || 'parse_err',
-                        graphqlName: (rc.operationName as any) || (rc.request.operationName as any),
-                    });
-
-                    rc.errors.forEach((error) => {
-                        if (error.path || error.name !== 'GraphQLError') {
-                            scope.setExtras({
-                                path: error.path,
-                            });
-                            Sentry.captureException(error);
-                        } else {
-                            scope.setExtras({});
-                            Sentry.captureMessage(`GraphQLWrongQuery: ${error.message}`);
-                        }
-                    });
-                });
-            },
-        };
+// Create the server app with its router/log/session and error management
+const app = new Koa();
+app.keys = [ process.env.COOKIE_SIGNING_KEY ]; // random key used to sign cookies
+app
+  .use(async (ctx, next) => {
+    const start = Date.now();
+    await next();
+    logger.info(`${ctx.method} ${ctx.url} - ${Date.now() - start}ms`);
+  })
+  .use(session({
+    store: {
+      get: key => dao.getSession(key),
+      set: (key, sess) => dao.setSession(key, sess),
+      destroy: key => dao.deleteSession(key)
     },
-} : {};
+    maxAge: 3600000 // 1h
+  }, app))
+  .use(router.routes())
+  .use(router.allowedMethods())
+  .on('error', err => {
+    logger.error(err.stack);
+    Sentry.captureException(err);
+  });
 
-const server = new ApolloServer({
-    typeDefs, resolvers, context, plugins: [apolloServerSentryPlugin], introspection: false, playground: false, debug: false
-});
-
-
+logger.info('Connecting to the databases...')
 dao.load()
-    .then(() => server.listen())
-    .then(({ url }: { url: string }) => {
-        logger.info(`🚀 Server ready at ${url}`);
-    })
-    .catch(error => Sentry.captureException(error));
+  .then(() => {
+    app.listen(3000)
+    logger.info(`🚀 Server ready at http://localhost:3000`);
+  })
